added file creation tests to test_data_viz.py, finalized project

+++ data_viz.py
@@ -1,13 +1,29 @@
-import sys
-import matplotlib
-matplotlib.use('Agg')
+"""Functions for data visualization
+
+Parameters
+----------
+L : a list containing only integers and
+    floats that will be the data to be visualized
+out_file_name : the name of the output data visualization figure
+
+Returns
+-------
+boxplot : Creates a boxplot with the data in list L
+histogram : Creates a histogram with the data in list L
+combo: Creates a both a boxplot and histogram with the data in list L
+"""
+
 import matplotlib.pyplot as plt
 import math_lib
+import os.path
+import matplotlib
+matplotlib.use('Agg')
 
 
+# method for creating boxplots
 def boxplot(L, out_file_name):
 
-
+    # ensures correct input type
     if L is None:
         return None
     if len(L) == 0:
@@ -18,16 +34,24 @@ def boxplot(L, out_file_name):
     for l in L:
         if not isinstance(l, int) and not isinstance(l, float):
             raise ValueError('Invalid type in list.')
+    if os.path.exists(out_file_name) is True:
+        print('This file name already exists')
+        return
 
     out_file = out_file_name
-    plt.title('mean: ' + str(math_lib.list_mean(L)) + '  ' + 'stdev: ' + str(math_lib.list_stdev(L)), fontsize=20)
+    plt.title('mean: ' + str(math_lib.list_mean(L))
+              + '  ' + 'stdev: ' + str(math_lib.list_stdev(L)), fontsize=20)
     plt.xlabel('sample')
     plt.ylabel('values')
 
     plt.boxplot(L)
     plt.savefig(out_file, bbox_inches='tight')
 
+
+# method for creating histograms
 def histogram(L, out_file_name):
+
+    # ensures correct input type
     if L is None:
         return None
     if len(L) == 0:
@@ -38,16 +62,24 @@ def histogram(L, out_file_name):
     for l in L:
         if not isinstance(l, int) and not isinstance(l, float):
             raise ValueError('Invalid type in list.')
+    if os.path.exists(out_file_name) is True:
+        print('This file name already exists')
+        return
 
     out_file = out_file_name
-    plt.title('mean: ' + str(math_lib.list_mean(L)) + '  ' + 'stdev: ' + str(math_lib.list_stdev(L)), fontsize=20)
+    plt.title('mean: ' + str(math_lib.list_mean(L))
+              + '  ' + 'stdev: ' + str(math_lib.list_stdev(L)), fontsize=20)
     plt.xlabel('frequency')
     plt.ylabel('values')
 
     plt.hist(L)
     plt.savefig(out_file, bbox_inches='tight')
 
+
+# method for creating boxplots and histograms at the same time
 def combo(L, out_file_name):
+
+    # ensures correct input type
     if L is None:
         return None
     if len(L) == 0:
@@ -58,6 +90,9 @@ def combo(L, out_file_name):
     for l in L:
         if not isinstance(l, int) and not isinstance(l, float):
             raise ValueError('Invalid type in list.')
+    if os.path.exists(out_file_name) is True:
+        print('This file name already exists')
+        return
 
     out_file = out_file_name
     width = 10
@@ -71,5 +106,6 @@ def combo(L, out_file_name):
     ax.hist(L)
     ax.set_xlabel('frequency')
     ax.set_ylabel('values')
-    fig.suptitle('mean: ' + str(math_lib.list_mean(L)) + '  ' + 'stdev: ' + str(math_lib.list_stdev(L)), fontsize=20)
+    fig.suptitle('mean: ' + str(math_lib.list_mean(L)) + '  '
+                 + 'stdev: ' + str(math_lib.list_stdev(L)), fontsize=20)
     plt.savefig(out_file, bbox_inches='tight')
